feat: useSWR for groups
