+++ agir/groups/components/groupsPage/index.js
@@ -2,6 +2,6 @@ import { renderWithContext } from "@agir/lib/utils/react";
 
 renderWithContext(
   import("./GroupsPage").then((module) => module.default),
-  "mes-groupes",
+  "exportedContent",
   "mainApp"
 );
