+++ agir/groups/urls.py
@@ -102,4 +102,5 @@ urlpatterns = [
         views.InvitationAbuseReportingView.as_view(),
         name="report_invitation_abuse",
     ),
+    path("api/groupes/", views.UserGroupsView.as_view(), name="api_user_groups"),
 ]
